requirements.txt
psutil==5.7.3
pygame==2.0.0
py-cpuinfo==7.0.0
texttable
python-nmap
